+++ src/main/java/fr/eql/ai113/LesVentsDalizes/repository/CustomerDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerDao extends JpaRepository<Customer, Integer> {
+public interface CustomerDao extends JpaRepository<Customer, Long> {
 
 
     Customer findCustomerByEmail(String email); //existsByEmail(String email);
@@ -16,5 +16,5 @@ public interface CustomerDao extends JpaRepository<Customer, Integer> {
 
 
 
-    Customer findCustomerById(Integer id);
+    Customer findCustomerById(Long id);
 }
