+++ src/main/java/fr/eql/ai113/LesVentsDalizes/service/RegistrationService.java
@@ -21,8 +21,7 @@ public interface RegistrationService {
 
    Address checkIfAddressAlreadyUsed(Address address);
 
-    Customer findCustomerById(int id);
-
+    Customer findCustomerById(Long id);
 
 
 }
